feat: add internal API error parsing
